refactor(ValidationBundle): extract validation flow into ValidationBundle
